Refactor *_processor -> processing.*
