+++ medusa/__main__.py
@@ -61,21 +61,25 @@ import threading
 import time
 
 from configobj import ConfigObj
+from six import text_type
 
 from medusa import (
-    app, auto_post_processor, cache, db, event_queue, exception_handler,
+    app, cache, db, event_queue, exception_handler,
     helpers, metadata, name_cache, naming, network_timezones, providers,
-    scheduler, show_queue, show_updater, subtitles, torrent_checker, trakt_checker, version_checker
+    scheduler, show_queue, show_updater, subtitles, torrent_checker,
+    trakt_checker, version_checker
 )
 from medusa.common import SD, SKIPPED, WANTED
 from medusa.config import (
-    CheckSection, ConfigMigrator, check_setting_bool, check_setting_float, check_setting_int, check_setting_list,
+    CheckSection, ConfigMigrator, check_setting_bool, check_setting_float,
+    check_setting_int, check_setting_list,
     check_setting_str, load_provider_setting, save_provider_setting
 )
 from medusa.databases import cache_db, failed_db, main_db
 from medusa.event_queue import Events
 from medusa.indexers.config import INDEXER_TVDBV2, INDEXER_TVMAZE
 from medusa.logger.adapters.style import BraceAdapter
+from medusa.processing import auto
 from medusa.providers.generic_provider import GenericProvider
 from medusa.providers.nzb.newznab import NewznabProvider
 from medusa.providers.torrent.rss.rsstorrent import TorrentRssProvider
@@ -87,8 +91,6 @@ from medusa.server.core import AppWebServer
 from medusa.system.shutdown import Shutdown
 from medusa.themes import read_themes
 from medusa.tv import Series
-
-from six import text_type
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
@@ -1179,11 +1181,12 @@ class Application(object):
 
             # processors
             update_interval = datetime.timedelta(minutes=app.AUTOPOSTPROCESSOR_FREQUENCY)
-            app.auto_post_processor_scheduler = scheduler.Scheduler(auto_post_processor.PostProcessor(),
-                                                                    cycleTime=update_interval,
-                                                                    threadName='POSTPROCESSOR',
-                                                                    silent=not app.PROCESS_AUTOMATICALLY,
-                                                                    run_delay=update_interval)
+            app.auto_post_processor_scheduler = scheduler.Scheduler(
+                auto.PostProcessor(),
+                cycleTime=update_interval,
+                threadName='POSTPROCESSOR',
+                silent=not app.PROCESS_AUTOMATICALLY,
+                run_delay=update_interval)
             update_interval = datetime.timedelta(minutes=5)
             app.trakt_checker_scheduler = scheduler.Scheduler(trakt_checker.TraktChecker(),
                                                               cycleTime=datetime.timedelta(hours=1),
