+++ medusa/processing/auto.py
@@ -21,7 +21,7 @@ from __future__ import unicode_literals
 import logging
 import os.path
 
-from medusa import app
+from medusa import app, process_tv
 
 
 logger = logging.getLogger(__name__)
@@ -40,7 +40,6 @@ class PostProcessor(object):
         :param force: Forces postprocessing run
         :type force: bool
         """
-        import process_tv
         self.amActive = True
         try:
             if not os.path.isdir(app.TV_DOWNLOAD_DIR):
